XSPF: kill a warning

+++ modules/demux/playlist/xspf.c
@@ -436,7 +436,7 @@ static bool parse_track_node COMPLEX_INTERFACE
                 free(psz_uri);
 
             if (p_sys->i_track_id < 0
-             || p_sys->i_track_id >= (SIZE_MAX / sizeof(p_new_input)))
+             || (unsigned)p_sys->i_track_id >= (SIZE_MAX / sizeof(p_new_input)))
             {
                 input_item_node_AppendNode(p_input_node, p_new_node);
                 vlc_gc_decref(p_new_input);
